refactor: organize SLAM runtime as system
