+++ crates/kornia-slam/src/initialization/inertial_factor.rs
