+++ crates/kornia-slam/src/initialization/mod.rs
@@ -1,0 +1,7 @@
+//! Visual and inertial initialization for the SLAM system.
+
+pub mod imu;
+pub mod inertial_factor;
+pub mod two_view;
+
+pub use imu::{ImuInitConfig, ImuInitResult, ImuInitializer};
