+++ crates/kornia-slam/src/tracking/mod.rs
@@ -1,0 +1,9 @@
+//! Frame-to-map tracking, motion propagation, and tracking policies.
+
+pub mod optical_flow;
+mod policy;
+pub mod pose_estimation;
+mod state;
+
+pub use policy::{KeyframePolicy, TrackingLossRecoveryPolicy};
+pub use state::{SystemMode, SystemState, TrackingResult, TrackingStatus};
