+++ crates/kornia-slam/src/tracking/pose_estimation/map_projection/keypoint_grid.rs
