+++ crates/kornia-slam/src/tracking/pose_estimation/mod.rs
@@ -1,20 +1,11 @@
-//! Pose estimation algorithms and estimator modules.
+//! Pose estimation used by frame-to-map tracking.
 
-pub mod imu_init;
-pub mod inertial_init_factor;
 pub mod map_projection;
-pub mod optical_flow;
 pub mod pnp;
-pub mod two_view;
 
 use kornia_3d::pose::Pose3d;
 
-pub use imu_init::{ImuInitConfig, ImuInitResult, ImuInitializer};
 pub use map_projection::MapProjectionEstimator;
-pub use optical_flow::{
-    FlowSurvivor, KeypointCorrespondence, KltTracker, MapKeypointMatch, SurvivorFilterConfig,
-    Track, TrackId, TrackSet, TrackSetError,
-};
 
 /// Successful pose estimate returned by any estimator.
 #[derive(Debug, Clone)]
